+++ nextcord/ext/menus/constants.py
@@ -1,0 +1,8 @@
+import logging
+
+
+# consistency with the `nextcord` namespaced logging
+log = logging.getLogger(__name__)
+
+# default timeout parameter for menus in seconds
+DEFAULT_TIMEOUT = 180.0
