+++ nextcord/nextcord
@@ -1,0 +1,1 @@
+../nextcord
